% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_efftox.R
\name{simulate_efftox}
\alias{simulate_efftox}
\title{Simulate Operating Characteristics using EffTox}
\usage{
simulate_efftox(
  ndose,
  ssizerange,
  target_t,
  target_e,
  startdose = 1,
  ntrial = 10000,
  utilitytype = 1,
  prob = NULL,
  save_dir = ".",
  save_folder = "efftox_simulations",
  save_file = "efftox_simulation.csv"
)
}
\arguments{
\item{ndose}{Integer. Number of dose levels. (\strong{Required})}

\item{ssizerange}{Integer vector. Range of number of cohorts to simulate. (\strong{Required})}

\item{target_t}{Numeric. Target toxicity probability. (\strong{Required})}

\item{startdose}{Integer. Starting dose level. (Default is \code{1})}

\item{ntrial}{Integer. Number of random trial replications. (Default is \code{10000})}

\item{utilitytype}{Integer. Type of utility structure. (Default is \code{1})
\itemize{
\item If set to \code{1}: Use preset weights (w11 = 0.6, w00 = 0.4)
\item If set to \code{2}: Use (w11 = 1, w00 = 0)
}}

\item{prob}{Fixed probability vectors. If not specified, a random scenario is used by default.
Use this parameter to provide fixed probability vectors as a list of the following named elements:
\itemize{
\item \code{pE}: Numeric vector of efficacy probabilities for each dose level.
\item \code{pT}: Numeric vector of toxicity probabilities for each dose level.
\item \code{obd}: Integer indicating the index of the true Optimal Biological Dose (OBD).
\item \code{mtd}: Integer indicating the index of the true Maximum Tolerated Dose (MTD).
}

For example:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{prob <- list(
  pE = c(0.4, 0.5, 0.6, 0.6, 0.6),
  pT = c(0.1, 0.2, 0.3, 0.4, 0.4),
  obd = 3,
  mtd = 2
)
``
}\if{html}{\out{</div>}}}

\item{save_dir}{Directory to save output folders. Default is (\code{"."}).}

\item{save_folder}{Folder name. (Default is "boin12_simulations")}

\item{save_file}{File name. (Default is "boin12_simulation.csv")}

\item{lower_e}{Numeric. Minimum acceptable efficacy probability. (\strong{Required})}
}
\value{
Results are saved as CSV files organized by OBD within folders.
}
\description{
This function runs simulations of the EffTox design by
evaluating operating characteristics over a range of cohort sizes. For each
dose level within the user-specified range, it performs multiple trials and saves the results to a corresponding file.
}
